+++ gis4wrf/plugin/geo.py
@@ -35,7 +35,7 @@ def rect_to_bbox(rect: QgsRectangle) -> BoundingBox2D:
 
 def update_domain_outline_layers(canvas: QgsMapCanvas, project: gis4wrf.core.Project,
                                  zoom_out=True) -> None:
-    gdal_ds = gis4wrf.core.convert_project_to_ogr_outlines(project)
+    gdal_ds = gis4wrf.core.convert_project_to_gdal_outlines(project)
     gdal_layer = gdal_ds.GetLayer(0) # type: ogr.Layer
     gdal_srs = gdal_layer.GetSpatialRef() # type: osr.SpatialReference
     proj4 = gdal_srs.ExportToProj4()
