+++ gis4wrf/plugin/ui/helpers.py
@@ -29,7 +29,7 @@ from PyQt5.QtWebKitWidgets import QWebView
 from qgis.gui import QgisInterface
 from qgis.core import QgsMapLayer
 
-# NOTE: Do not import anything from gis4wrf.core here.
+# NOTE: Do not import anything from gis4wrf.core or other gis4wrf.plugin module depending on core here.
 #       The helpers module is used in the bootstrapping UI.
 
 from gis4wrf.plugin.constants import PLUGIN_NAME
@@ -261,6 +261,9 @@ def wrap_error(parent, fn):
         return
 
 def dispose_after_delete(layer: QgsMapLayer, dispose: Callable[[],None]) -> None:
+    # Lazy import to work around restriction explained at top of this file.
+    from gis4wrf.plugin.ui.thread import TaskThread
+
     # There is no signal indicating that the layer has been fully removed.
     # Therefore in the willBeDeleted signal we need to give control back to
     # the main event loop and run the dispose operation asynchronously. Note that
