+++ gis4wrf/plugin/ui/widget_met.py
@@ -14,11 +14,11 @@ from PyQt5.QtWidgets import (
 from gis4wrf.core import (
     met_datasets, get_met_products, is_met_dataset_downloaded, get_met_dataset_path, download_met_dataset,
     CRS)
-from gis4wrf.plugin.ui.helpers import add_grid_lineedit, MessageBar, reraise
-from gis4wrf.plugin.ui.thread import TaskThread
 from gis4wrf.plugin.options import get_options
 from gis4wrf.plugin.geo import rect_to_bbox
-from .broadcast import Broadcast
+from gis4wrf.plugin.ui.helpers import add_grid_lineedit, MessageBar, reraise
+from gis4wrf.plugin.ui.thread import TaskThread
+from gis4wrf.plugin.ui.broadcast import Broadcast
 
 
 DECIMALS = 50
